requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
xhtml2pdf==0.2.17  
reportlab==4.4.4   
Pillow==12.0.0     
